requirements.txt
python-dotenv==0.21.0
influxdb-client==1.32.0
pyserial==3.5
minimalmodbus==2.0.1
